refactor(plan-202): move host-agent control protocol to mvm-core (1c-prep)
The RegisterVm/DeregisterVm/SignedControl/ControlResponse wire contract was in
mvm-hostd (1a), but the sender of these messages is mvm-backend (the up/start
path), which by the architecture cannot depend on mvm-hostd. Rather than
duplicate the wire shape as raw JSON in mvm-backend (the SubprocessConfig
pattern), move the protocol to mvm-core — where the ServiceCall/broker types
already live — so the daemon (mvm-hostd) and the registering backend
(mvm-backend) share one set of types.

mvm-hostd::broker::control becomes a thin re-export, so the daemon's
super::control::… paths and its tests are unchanged. Adds serde_jcs to mvm-core
(sync; check-core-runtime-free stays green — no tokio pulled). Pure refactor,
no behavior change; 8 control tests + 4 daemon tests green.
