+++ crates/mvm-core/Cargo.toml
@@ -11,6 +11,7 @@ authors.workspace = true
 [dependencies]
 anyhow.workspace = true
 base64.workspace = true
+serde_jcs = "0.1"
 chrono.workspace = true
 # `protocol::signed_config` verifies
 # subprocess startup config envelopes via Ed25519. The verify path
